Extract transform/camera dialog HTML into engine/src/ui/dialogs.ts
Six show*Dialog methods (translation, quaternion, angle-axis, camera
position, camera rotation, rotation center) each rebuilt an identical
~25-line modal overlay + close/cancel/escape-key wiring around a
different HTML template. Pulled the shared shell into
createModalDialog() and each template into its own exported function,
so main.ts keeps only the apply-button logic that actually touches
instance state (camera, controls, transformation matrices).

Also moved escapeHtml and addTooltipsToTruncatedFilenames - both
already state-free - into the same module.

main.ts: 14,928 -> 14,511 lines. Updated CLAUDE.md's project-structure
tree to reflect that ui/dialogs.ts and utils/matrix.ts are real now,
not proposed.

Verified: engine tsc clean, full webpack compile clean, all 59
extension tests passing under xvfb-run.
